+++ CHANGELOG.md
@@ -182,4 +182,7 @@ Added the `fate` command in `fun.fun`. Also changed the 8ball and commit command
 Added `util.qalc`, adding the `qalc` command for evaluating mathematical expressions with libqalculate. Resolves issue #88.
 
 #### 1.22.1
-Removed unclear responses for the 8ball command in `fun.fun`
+Removed unclear responses for the 8ball command in `fun.fun`.
+
+### 1.23.0
+Added `util.steam`, which has the `steam` command for looking up video games on Steam via the Steam API.